// web/src/api/generated/models/RulePreferenceView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RuleParametersView } from './RuleParametersView';
import {
    RuleParametersViewFromJSON,
    RuleParametersViewFromJSONTyped,
    RuleParametersViewToJSON,
    RuleParametersViewToJSONTyped,
} from './RuleParametersView';

/**
 * 
 * @export
 * @interface RulePreferenceView
 */
export interface RulePreferenceView {
    /**
     * 
     * @type {boolean}
     * @memberof RulePreferenceView
     */
    enabled: boolean;
    /**
     * 
     * @type {RuleParametersView}
     * @memberof RulePreferenceView
     */
    parameters: RuleParametersView;
    /**
     * 
     * @type {string}
     * @memberof RulePreferenceView
     */
    ruleKey: string;
}

/**
 * Check if a given object implements the RulePreferenceView interface.
 */
export function instanceOfRulePreferenceView(value: object): value is RulePreferenceView {
    if (!('enabled' in value) || value['enabled'] === undefined) return false;
    if (!('parameters' in value) || value['parameters'] === undefined) return false;
    if (!('ruleKey' in value) || value['ruleKey'] === undefined) return false;
    return true;
}

export function RulePreferenceViewFromJSON(json: any): RulePreferenceView {
    return RulePreferenceViewFromJSONTyped(json, false);
}

export function RulePreferenceViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): RulePreferenceView {
    if (json == null) {
        return json;
    }
    return {
        
        'enabled': json['enabled'],
        'parameters': RuleParametersViewFromJSON(json['parameters']),
        'ruleKey': json['ruleKey'],
    };
}

export function RulePreferenceViewToJSON(json: any): RulePreferenceView {
    return RulePreferenceViewToJSONTyped(json, false);
}

export function RulePreferenceViewToJSONTyped(value?: RulePreferenceView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'enabled': value['enabled'],
        'parameters': RuleParametersViewToJSON(value['parameters']),
        'ruleKey': value['ruleKey'],
    };
}